old-tests/submission/Microsoft/video/video_011.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: The 'src' IDL attribute of media elements reflects the content attribute (relative URI).</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#attr-media-src" />
  <meta name="assert" content="The 'src' IDL attribute of media elements reflects the content attribute." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
  <script type="text/javascript" src="../common/media.js"></script>
  
</head>
<body>
<pre>
The 'src' IDL attribute of media elements reflects the content attribute (relative URI).
</pre>
<pre>

http://www.w3.org/TR/html5/video.html#attr-media-src

The src IDL attribute on media elements must reflect the content 
attribute of the same name.

http://www.w3.org/TR/html5/infrastructure.html#reflect

If a reflecting IDL attribute is a DOMString attribute whose content 
attribute is defined to contain a URL, then on getting, the IDL 
attribute must resolve the value of the content attribute relative to 
the element and return the resulting absolute URL if that was 
successful, or the empty string otherwise; and on setting, must set the 
content attribute to the specified literal value. 

</pre>
<div id='log'></div>

<video id="video0" src="broken-uri">Your browser does not support media elements.</video>
<script type="text/javascript">    
        test(function()
        {
           assert_not_equals(document.getElementById("video0").src.match(/\/broken-uri$/), null);
        }, "The 'src' IDL attribute of media elements reflects the content attribute.");
    </script>
</body>
</html>
